move parity models to test folder
