+++ benchmarks/test_parity.py
@@ -5,8 +5,8 @@ import pytest
 import torch
 
 import tests.base.utils as tutils
-from benchmarks.parity_modules import ParityModuleRNN, ParityModuleMNIST
 from pytorch_lightning import Trainer, seed_everything
+from tests.base.models import ParityModuleRNN, ParityModuleMNIST
 
 
 @pytest.mark.parametrize('cls_model,max_diff', [
